Get nice feedback with RuleTester using AVA
