+++ test/no-invalid-end.js
@@ -1,8 +1,8 @@
 import test from 'ava';
-import {RuleTester} from 'eslint';
+import avaRuleTester from 'eslint-ava-rule-tester';
 import rule from '../rules/no-invalid-end';
 
-const ruleTester = new RuleTester({
+const ruleTester = avaRuleTester(test, {
 	env: {
 		es6: true
 	}
@@ -11,32 +11,30 @@ const ruleTester = new RuleTester({
 const errors = [{ruleId: 'no-invalid-end'}];
 const header = `const test = require('ava');\n`;
 
-test(() => {
-	ruleTester.run('no-invalid-end', rule, {
-		valid: [
-			header + 'test(t => {});',
-			header + 'test(t => { t.is(1, 1); });',
-			header + 'test.only(t => {});',
-			header + 'test.cb(t => { t.end(); });',
-			header + 'test.cb(t => { t.end.skip(); });',
-			header + 'test.cb.only(t => { t.end(); });',
-			header + 'notTest(t => { t.end(); });',
-			// shouldn't be triggered since it's not a test file
-			'test(t => { t.end(); });'
-		],
-		invalid: [
-			{
-				code: header + 'test(t => { t.end(); });',
-				errors
-			},
-			{
-				code: header + 'test.only(t => { t.end(); });',
-				errors
-			},
-			{
-				code: header + 'test(t => { t.end.skip(); });',
-				errors
-			}
-		]
-	});
+ruleTester.run('no-invalid-end', rule, {
+	valid: [
+		header + 'test(t => {});',
+		header + 'test(t => { t.is(1, 1); });',
+		header + 'test.only(t => {});',
+		header + 'test.cb(t => { t.end(); });',
+		header + 'test.cb(t => { t.end.skip(); });',
+		header + 'test.cb.only(t => { t.end(); });',
+		header + 'notTest(t => { t.end(); });',
+		// shouldn't be triggered since it's not a test file
+		'test(t => { t.end(); });'
+	],
+	invalid: [
+		{
+			code: header + 'test(t => { t.end(); });',
+			errors
+		},
+		{
+			code: header + 'test.only(t => { t.end(); });',
+			errors
+		},
+		{
+			code: header + 'test(t => { t.end.skip(); });',
+			errors
+		}
+	]
 });
