+++ test/no-only-test.js
@@ -1,8 +1,8 @@
 import test from 'ava';
-import {RuleTester} from 'eslint';
+import avaRuleTester from 'eslint-ava-rule-tester';
 import rule from '../rules/no-only-test';
 
-const ruleTester = new RuleTester({
+const ruleTester = avaRuleTester(test, {
 	env: {
 		es6: true
 	}
@@ -11,29 +11,27 @@ const ruleTester = new RuleTester({
 const errors = [{ruleId: 'no-only-test'}];
 const header = `const test = require('ava');\n`;
 
-test(() => {
-	ruleTester.run('no-only-test', rule, {
-		valid: [
-			header + 'test("my test name", t => { t.pass(); });',
-			header + 'test.cb("my test name", t => { t.pass(); t.end(); });',
-			header + 'test(t => { t.pass(); }); test(t => { t.pass(); });',
-			header + 'notTest.only();',
-			// shouldn't be triggered since it's not a test file
-			'test.only(t => {});'
-		],
-		invalid: [
-			{
-				code: header + 'test.only(t => { t.pass(); });',
-				errors
-			},
-			{
-				code: header + 'test.cb.only(t => { t.pass(); t.end(); });',
-				errors
-			},
-			{
-				code: header + 'test.only.cb(t => { t.pass(); t.end(); });',
-				errors
-			}
-		]
-	});
+ruleTester.run('no-only-test', rule, {
+	valid: [
+		header + 'test("my test name", t => { t.pass(); });',
+		header + 'test.cb("my test name", t => { t.pass(); t.end(); });',
+		header + 'test(t => { t.pass(); }); test(t => { t.pass(); });',
+		header + 'notTest.only();',
+		// shouldn't be triggered since it's not a test file
+		'test.only(t => {});'
+	],
+	invalid: [
+		{
+			code: header + 'test.only(t => { t.pass(); });',
+			errors
+		},
+		{
+			code: header + 'test.cb.only(t => { t.pass(); t.end(); });',
+			errors
+		},
+		{
+			code: header + 'test.only.cb(t => { t.pass(); t.end(); });',
+			errors
+		}
+	]
 });
