+++ test/no-statement-after-end.js
@@ -1,8 +1,8 @@
 import test from 'ava';
-import {RuleTester} from 'eslint';
+import avaRuleTester from 'eslint-ava-rule-tester';
 import rule from '../rules/no-statement-after-end';
 
-const ruleTester = new RuleTester({
+const ruleTester = avaRuleTester(test, {
 	env: {
 		es6: true
 	}
@@ -21,35 +21,33 @@ function cbTest(contents, prependHeader) {
 	return ret;
 }
 
-test(() => {
-	ruleTester.run('no-statement-after-end', rule, {
-		valid: [
-			cbTest('t.end();'),
-			cbTest('t.is(1, 1); t.end();'),
-			cbTest('notT.end(); t.is(1, 1);'),
-			cbTest('if (t.context.a === 1) { return t.end(); } \n t.is(1, 1); t.end();'),
-			cbTest('return t.end();'),
-			cbTest('t.end(); return;'),
+ruleTester.run('no-statement-after-end', rule, {
+	valid: [
+		cbTest('t.end();'),
+		cbTest('t.is(1, 1); t.end();'),
+		cbTest('notT.end(); t.is(1, 1);'),
+		cbTest('if (t.context.a === 1) { return t.end(); } \n t.is(1, 1); t.end();'),
+		cbTest('return t.end();'),
+		cbTest('t.end(); return;'),
 			// valid because it is not a test file (no header)
-			cbTest('t.end(); t.is(1, 1);', false)
-		],
-		invalid: [
-			{
-				code: cbTest('t.end(); t.is(1, 1);'),
-				errors
-			},
-			{
-				code: cbTest('t.end(); return 3 + 4;'),
-				errors
-			},
-			{
-				code: cbTest('t.end(); console.log("end");'),
-				errors
-			},
-			{
-				code: cbTest('if (t.context.a === 1) { t.end(); }\nt.is(1, 1); t.end();'),
-				errors
-			}
-		]
-	});
+		cbTest('t.end(); t.is(1, 1);', false)
+	],
+	invalid: [
+		{
+			code: cbTest('t.end(); t.is(1, 1);'),
+			errors
+		},
+		{
+			code: cbTest('t.end(); return 3 + 4;'),
+			errors
+		},
+		{
+			code: cbTest('t.end(); console.log("end");'),
+			errors
+		},
+		{
+			code: cbTest('if (t.context.a === 1) { t.end(); }\nt.is(1, 1); t.end();'),
+			errors
+		}
+	]
 });
