+++ test/no-unknown-modifiers.js
@@ -1,8 +1,8 @@
 import test from 'ava';
-import {RuleTester} from 'eslint';
+import avaRuleTester from 'eslint-ava-rule-tester';
 import rule from '../rules/no-unknown-modifiers';
 
-const ruleTester = new RuleTester({
+const ruleTester = avaRuleTester(test, {
 	env: {
 		es6: true
 	}
@@ -11,75 +11,73 @@ const ruleTester = new RuleTester({
 const ruleError = {ruleId: 'no-unknown-modifiers'};
 const header = `const test = require('ava');\n`;
 
-test(() => {
-	ruleTester.run('no-unknown-modifiers', rule, {
-		valid: [
-			`${header} test(t => {});`,
-			`${header} test.after(t => {});`,
-			`${header} test.afterEach(t => {});`,
-			`${header} test.before(t => {});`,
-			`${header} test.beforeEach(t => {});`,
-			`${header} test.cb(t => {});`,
-			`${header} test.cb.only(t => {});`,
-			`${header} test.only(t => {});`,
-			`${header} test.serial(t => {});`,
-			`${header} test.skip(t => {});`,
-			`${header} test.todo(t => {});`,
-			`${header} test.after.always(t => {});`,
-			`${header} test.afterEach.always(t => {});`,
-			`${header} test.failing(t => {});`,
-			// shouldn't be triggered since it's not a test file
-			`test.foo(t => {});`
-		],
-		invalid: [
-			{
-				code: `${header} test.foo(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `foo`.'}
-				]
-			},
-			{
-				code: `${header} test.onlu(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `onlu`.'}
-				]
-			},
-			{
-				code: `${header} test.beforeeach(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `beforeeach`.'}
-				]
-			},
-			{
-				code: `${header} test.c.only(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `c`.'}
-				]
-			},
-			{
-				code: `${header} test.cb.onlu(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `onlu`.'}
-				]
-			},
-			{
-				code: `${header} test.foo.bar.baz(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `foo`.'}
-				]
-			},
-			{
-				code: `${header} test.default(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `default`.'}
-				]
-			},
-			{
-				code: `${header} test.test(t => {});`,
-				errors: [
-					{...ruleError, message: 'Unknown test modifier `test`.'}
-				]
-			}
-		]
-	});
+ruleTester.run('no-unknown-modifiers', rule, {
+	valid: [
+		`${header} test(t => {});`,
+		`${header} test.after(t => {});`,
+		`${header} test.afterEach(t => {});`,
+		`${header} test.before(t => {});`,
+		`${header} test.beforeEach(t => {});`,
+		`${header} test.cb(t => {});`,
+		`${header} test.cb.only(t => {});`,
+		`${header} test.only(t => {});`,
+		`${header} test.serial(t => {});`,
+		`${header} test.skip(t => {});`,
+		`${header} test.todo(t => {});`,
+		`${header} test.after.always(t => {});`,
+		`${header} test.afterEach.always(t => {});`,
+		`${header} test.failing(t => {});`,
+		// shouldn't be triggered since it's not a test file
+		`test.foo(t => {});`
+	],
+	invalid: [
+		{
+			code: `${header} test.foo(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `foo`.'}
+			]
+		},
+		{
+			code: `${header} test.onlu(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `onlu`.'}
+			]
+		},
+		{
+			code: `${header} test.beforeeach(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `beforeeach`.'}
+			]
+		},
+		{
+			code: `${header} test.c.only(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `c`.'}
+			]
+		},
+		{
+			code: `${header} test.cb.onlu(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `onlu`.'}
+			]
+		},
+		{
+			code: `${header} test.foo.bar.baz(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `foo`.'}
+			]
+		},
+		{
+			code: `${header} test.default(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `default`.'}
+			]
+		},
+		{
+			code: `${header} test.test(t => {});`,
+			errors: [
+				{...ruleError, message: 'Unknown test modifier `test`.'}
+			]
+		}
+	]
 });
